:up: Krews update Done :"Refactored string comparisons to use equalsIgnoreCase, replaced valueOf with parseInt for improved efficiency"

+++ Emulator/src/main/java/com/eu/habbo/habbohotel/commands/ShutdownCommand.java
@@ -44,7 +44,7 @@ public class ShutdownCommand extends Command {
         }
         RoomTrade.TRADING_ENABLED = false;
         ShutdownEmulator.timestamp = Emulator.getIntUnixTimestamp() + (60 * minutes);
-        Emulator.getThreading().run(new ShutdownEmulator(message), minutes * 60 * 1000);
+        Emulator.getThreading().run(new ShutdownEmulator(message), (long) minutes * 60 * 1000);
         return true;
     }
 }
